pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vpype-vfab"
version = "0.4.1"
description = "vpype plugin for vfab plotter management integration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "bkuri", email = "ben@kurita.org"}
]
keywords = ["plotter", "vpype", "vfab", "axidraw", "generative-art"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Artistic",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Artistic Software",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "vpype>=1.14.0",
    "click>=8.0.0",
    "pydantic>=2.0",
    "platformdirs>=4.0.0",
    "pyyaml>=6.0.2",
    "websockets>=12.0",
    "fastapi>=0.104.0",
]

[project.optional-dependencies]
vfab = ["plotty>=1.1.0"]
monitoring = ["rich>=13.0.0"]
dev = [
    "pytest>=8.3.2",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.10.0",
    "ruff>=0.6.8",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/bkuri/vpype-vfab"
Documentation = "https://github.com/bkuri/vpype-vfab/docs"
Repository = "https://github.com/bkuri/vpype-vfab"
"Bug Tracker" = "https://github.com/bkuri/vpype-vfab/issues"

[project.entry-points."vpype.plugins"]
vfab-add = "vpype_vfab.commands:vfab_add"
vfab-queue = "vpype_vfab.commands:vfab_queue"
vfab-status = "vpype_vfab.commands:vfab_status"
vfab-list = "vpype_vfab.commands:vfab_list"
vfab-monitor = "vpype_vfab.commands:vfab_monitor"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=vpype_vfab --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.uv.workspace]
members = [
    "sandbox",
    "sandbox/plotty",
]
